// src/generated/models/Poc_productsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Poc_productspoc_category = {
  893960000: 'Peripherals',
  893960001: 'Monitors',
  893960002: 'Laptops',
  893960003: 'Accessories',
  893960004: 'Furniture',
  893960005: 'Other'
} as const;
export type Poc_productspoc_category = keyof typeof Poc_productspoc_category;
export const Poc_productsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Poc_productsstatecode = keyof typeof Poc_productsstatecode;
export const Poc_productsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Poc_productsstatuscode = keyof typeof Poc_productsstatuscode;

export interface Poc_productsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  poc_available?: boolean;
  poc_category?: Poc_productspoc_category;
  poc_description?: string;
  poc_image?: string;
  poc_product1?: string;
  poc_productid: string;
  poc_unitprice?: number;
  statecode: Poc_productsstatecode;
  statuscode?: Poc_productsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Poc_products extends Poc_productsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  poc_availablename?: string;
  poc_categoryname?: string;
  poc_unitprice_base?: number;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}